move generateCatalogDiffsAndDisableConnectionsIfNeeded to sourceHandler
